Stash docs progress
