+++ dev_requirements.txt
@@ -1,0 +1,1 @@
+pdoc3==0.10.0
